backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
numpy>=1.24,<2.0
opencv-python-headless>=4.8,<5.0
stripe>=7.0,<8.0
Pillow>=10.0,<11.0
# mediapipe installed separately - requires cmake and Python 3.10+
# pip install mediapipe>=0.10
